Refactor Classe and Sector services to use int for sector IDs; update related methods and DTOs for consistency

+++ metier/src/main/java/com/exam/metier/dao/IClasseDao.java
@@ -12,7 +12,7 @@ public interface IClasseDao {
 
     List<Classes> getAllClasses();
 
-    List<Classes> getClassesBySector(Long sectorId);
+    List<Classes> getClassesBySector(int sectorId);
 
     Classes getClasseById(int id);
 }
